# api/openapi.yaml
openapi: 3.1.0
...
info:
    title: Nexus Open API
    description: |
        REST API for Nexus Open - an open-source iCUE Nexus companion app.
        Provides endpoints for device control, configuration management, and zone/plugin interactions.
    contact:
        name: Team
        url: https://github.com/mantonx/nexus-open
    version: 2.0.0
paths:
    /api/config:
        get:
            tags:
                - Config
            summary: Get current configuration
            description: Returns the current application configuration including display settings, location, and modules
            operationId: getConfig
            requestBody:
                content:
                    '*/*': {}
            responses:
                "200":
                    description: Configuration retrieved successfully
                    content:
                        application/json:
                            schema:
                                properties:
                                    background_color:
                                        type: string
                                        description: Background color in hex format
                                        example: '#000000'
                                    background_image:
                                        type: string
                                        description: Background image filename
                                        example: background.png
                                    display:
                                        $ref: '#/components/schemas/Display'
                                    image_paths:
                                        type: array
                                        description: List of custom image paths
                                        items:
                                            type: string
                                    text_color:
                                        type: string
                                        description: Text color in hex format
                                        example: '#FFFFFF'
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
        post:
            tags:
                - Config
            summary: Update configuration
            description: Updates the application configuration with new settings
            operationId: updateConfig
            parameters:
                - name: config
                  in: body
                  description: Configuration object to update
                  required: true
                  schema:
                    type: Config
            requestBody:
                content:
                    application/json: {}
            responses:
                "200":
                    description: Configuration updated successfully
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {}
                                    message:
                                        type: string
                                        description: Human-readable success message
                                        example: Configuration updated successfully
                                    status:
                                        type: string
                                        description: Status of the operation
                                        example: success
                "400":
                    description: Invalid configuration provided
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
                "500":
                    description: Failed to save configuration
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
    /api/device/brightness:
        post:
            tags:
                - Device
            summary: Set display brightness
            description: Sets the brightness of the iCUE Nexus device display (0-100)
            operationId: setBrightness
            parameters:
                - name: brightness
                  in: body
                  description: Brightness level to set
                  required: true
                  schema:
                    type: BrightnessRequest
            requestBody:
                content:
                    application/json: {}
            responses:
                "200":
                    description: Brightness updated successfully
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {}
                                    message:
                                        type: string
                                        description: Human-readable success message
                                        example: Configuration updated successfully
                                    status:
                                        type: string
                                        description: Status of the operation
                                        example: success
                "400":
                    description: Invalid brightness value
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
                "503":
                    description: Device not available
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
    /api/device/info:
        get:
            tags:
                - Device
            summary: Get device information
            description: Returns information about the connected iCUE Nexus device including firmware version and hardware details
            operationId: getDeviceInfo
            requestBody:
                content:
                    '*/*': {}
            responses:
                "200":
                    description: Device information retrieved successfully
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {}
                                    message:
                                        type: string
                                        description: Human-readable success message
                                        example: Configuration updated successfully
                                    status:
                                        type: string
                                        description: Status of the operation
                                        example: success
                "503":
                    description: Device not available
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                        description: The error type/category
                                        example: Bad Request
                                    message:
                                        type: string
                                        description: Detailed error message
                                        example: Invalid brightness value provided
...
    /api/layout/draft:
        get:
            tags:
                - Layout
            summary: Get or open draft layout
            description: Returns the current draft, opening a new one from the committed config if none is active.
            operationId: getDraft
            responses:
                "200":
                    description: Draft layout
                    content:
                        application/json:
                            schema:
                                properties:
                                    active:
                                        type: boolean
                                    layout:
                                        type: object
                "503":
                    description: Draft manager not available
        put:
            tags:
                - Layout
            summary: Replace draft layout
            description: Replaces the entire draft with the provided config and previews it live on device.
            operationId: putDraft
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            description: Full zone.Config
            responses:
                "200":
                    description: Draft updated
                "422":
                    description: Invalid layout
    /api/layout/draft/zones:
        post:
            tags:
                - Layout
            summary: Add zone to draft
            description: Appends a zone to the specified page in the draft and redistributes widths.
            operationId: addDraftZone
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                page_index:
                                    type: integer
                                plugin:
                                    type: string
                                refresh_ms:
                                    type: integer
            responses:
                "200":
                    description: Zone added
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        properties:
                                            id:
                                                type: string
                "422":
                    description: Zone cap exceeded or width redistribution impossible
...
    /api/layout/commit:
        post:
            tags:
                - Layout
            summary: Commit draft to store
            description: Persists the current draft to the database. The device display is already showing the draft; commit just makes it durable.
            operationId: commitDraft
            responses:
                "200":
                    description: Draft committed
                "409":
                    description: No active draft
...
servers:
    - url: http://localhost:1985
...